Roomoval
The Android and Rust code have previously managed joint custody of the derived data database.  With more complex migrations now required, we need to make the Android side read-only.  To achieve that, the Android side will remove Room and instead rely on more primitive SQLite APIs for read-only access.

As part of implementing this change, database management throughout the SDK is being refactored.

There will be multiple representations of the data:
 - Database representation (and some Entity representations in the places that Room hasn’t been removed yet).  These representations are not as type safe and don’t match Kotlin best practices in all cases.
 - Once #615 is implemented there will also be network representations
 - Type safe models, which often appear in the public API  The Database and Network representations will be converted to and from the type safe model representation.

Still pending before merging:
 - Many database methods are stubbed out.  See DbDerivedDataRepository
 - Some database write operations are occurring and these should be replaced with Rust implementations.  Many of these are happening inside SdkSynchronizer
 - Unit testing of the Android database query implementations (given the Rust layer has been set up and known query results are expected back out)

+++ demo-app/src/main/java/cash/z/ecc/android/sdk/demoapp/demos/listutxos/ListUtxosFragment.kt
@@ -12,7 +12,6 @@ import cash.z.ecc.android.sdk.Initializer
 import cash.z.ecc.android.sdk.SdkSynchronizer
 import cash.z.ecc.android.sdk.Synchronizer
 import cash.z.ecc.android.sdk.block.CompactBlockProcessor
-import cash.z.ecc.android.sdk.db.entity.ConfirmedTransaction
 import cash.z.ecc.android.sdk.demoapp.BaseDemoFragment
 import cash.z.ecc.android.sdk.demoapp.databinding.FragmentListUtxosBinding
 import cash.z.ecc.android.sdk.demoapp.ext.requireApplicationContext
@@ -21,6 +20,7 @@ import cash.z.ecc.android.sdk.demoapp.util.mainActivity
 import cash.z.ecc.android.sdk.ext.collectWith
 import cash.z.ecc.android.sdk.internal.twig
 import cash.z.ecc.android.sdk.model.BlockHeight
+import cash.z.ecc.android.sdk.model.ConfirmedTransaction
 import cash.z.ecc.android.sdk.model.LightWalletEndpoint
 import cash.z.ecc.android.sdk.model.ZcashNetwork
 import cash.z.ecc.android.sdk.model.defaultForNetwork
@@ -149,7 +149,7 @@ class ListUtxosFragment : BaseDemoFragment<FragmentListUtxosBinding>() {
 
             lifecycleScope.launch {
                 withContext(Dispatchers.IO) {
-                    finalCount = (synchronizer as SdkSynchronizer).getTransactionCount()
+                    finalCount = (synchronizer as SdkSynchronizer).getTransactionCount().toInt()
                     withContext(Dispatchers.Main) {
                         @Suppress("MagicNumber")
                         delay(100)
@@ -199,7 +199,7 @@ class ListUtxosFragment : BaseDemoFragment<FragmentListUtxosBinding>() {
             lifecycleScope.launch {
                 withContext(Dispatchers.IO) {
                     synchronizer.prepare()
-                    initialCount = (synchronizer as SdkSynchronizer).getTransactionCount()
+                    initialCount = (synchronizer as SdkSynchronizer).getTransactionCount().toInt()
                 }
             }
             synchronizer.clearedTransactions.collectWith(lifecycleScope, ::onTransactionsUpdated)
